perf+fix: P1 gap closures from the audit
- dashboard::snapshot collapses four serial D1 round-trips into a single
  batched three-GROUP-BY + top-10 activity query. Server-side aggregation
  removes the Vec<Row> shuttle and a ~40-160 ms landing-page latency
  contribution.
- Cache-Control is method-aware: GET/HEAD render HTML ('private, no-cache',
  bfcache-eligible, revalidates), everything else stays 'no-store'.
  http::Method isn't structural-match-safe so use an if/else equality
  comparison rather than a match on associated constants.
- Wire types: StageCounts/ResourceStatusCounts/PlacementStatusCounts and
  the three *_total fields on DashboardSnapshot switch from usize to u32
  so the wasm32 client and x86_64 server agree on integer width across
  the hydration boundary. MetricCard follows.
- normalize_text caps on char count, not byte length — multibyte input
  was hitting the limit well below the documented 'N characters'.
- list_households / list_resources / list_placements all grow a hard
  LIMIT 500 so a runaway table can't ship unbounded JSON through the
  Worker.

Verified post-deploy v4 (a82a4a89-eefe-4364-9893-f80202ea9d00):
- GET / /households /inventory /placements /activity → 200,
  cache-control: private, no-cache.
- POST /api/* → cache-control: no-store.
- Dashboard SSR payload decodes cleanly: household_total=3,
  resource_total=4, placement_total=2, recent_activity=5 entries.
- All three hashed assets (js/wasm/css) resolve 200.

+++ src/components/dashboard_page.rs
@@ -100,7 +100,7 @@ fn DashboardBoard(data: DashboardSnapshot) -> impl IntoView {
 }
 
 #[component]
-fn MetricCard(#[prop(into)] label: String, value: usize) -> impl IntoView {
+fn MetricCard(#[prop(into)] label: String, value: u32) -> impl IntoView {
     view! {
         <article class="metric-card">
             <span class="metric-label">{label}</span>
